@@ -22,25 +22,17 @@
 		}
 	],
 	"require": {
-		"php": "^7.2",
-		"squizlabs/php_codesniffer": "^3.6.0",
-		"phpcompatibility/phpcompatibility-wp": "^2.1.1",
-		"wp-coding-standards/wpcs": "^2.3.0",
-		"phpcsstandards/phpcsutils": "^1.0",
-		"phpcsstandards/phpcsextra": "^1.0",
-		"slevomat/coding-standard": "^7.0"
+		"php": ">=7.4",
+		"phpcompatibility/phpcompatibility-wp": "^2.1.4",
+		"wp-coding-standards/wpcs": "dev-hotifx/escape-output-sniff",
+		"slevomat/coding-standard": "^8.13.0"
 	},
 	"require-dev": {
-		"phpunit/phpunit": "^4.0 || ^5.0 || ^6.0 || ^7.0",
-		"phpcsstandards/phpcsdevtools": "^1.0",
+		"phpunit/phpunit": "^7.0",
+		"phpcsstandards/phpcsdevtools": "^1.2.0",
 		"php-parallel-lint/php-parallel-lint": "^1.3.2",
 		"php-parallel-lint/php-console-highlighter": "^1.0.0",
 		"roave/security-advisories": "dev-master"
-	},
-	"autoload-dev": {
-		"psr-4": {
-			"EightshiftCS\\Eightshift\\": "Eightshift/"
-		}
 	},
 	"minimum-stability": "dev",
 	"prefer-stable": true,
