@@ -5,15 +5,14 @@
  *
  * @package EightshiftCS
  *
- * @author  Eightshift <team@eightshift.com>
+ * @author  Eightshift <team.wordpress@infinum.com>
  * @license https://opensource.org/licenses/MIT MIT
  * @link    https://github.com/infinum/eightshift-coding-standards
- *
- * @since 1.4.0
  */
 
 namespace EightshiftCS\Eightshift\Sniffs\Security;
 
+use PHP_CodeSniffer\Util\Tokens;
 use PHPCSUtils\Utils\GetTokensAsString;
 use PHPCSUtils\Utils\UseStatements;
 use WordPressCS\WordPress\Sniffs\Security\EscapeOutputSniff;
@@ -22,28 +21,45 @@ use WordPressCS\WordPress\Sniffs\Security\EscapeOutputSniff;
  * Override the WordPress.Security.EscapeOutput sniff
  *
  * This sniff will ignore escaping errors whenever it finds the
- * EightshiftLibs\Helpers\Components::render or
- * EightshiftLibs\Helpers\Components::outputCssVariables methods.
+ * EightshiftLibs\Helpers\Components::$allowedMethods, where the $allowedMethods are by default `render` and `outputCssVariables`, but can be extended in the ruleset.
  *
- * These methods are considered safe because the components
- * should be properly escaped on the output.
+ * $allowedMethods are considered safe.
+ *
+ * @since 2.0.0 Add list of allowed static methods that shouldn't trigger the sniff error.
+ * @since 1.4.0
  */
 class ComponentsEscapeSniff extends EscapeOutputSniff
 {
 	/**
-	 * A fully qualified class name for Components class override
+	 * A fully qualified class name for Components class override.
 	 *
 	 * You should put the fully qualified class name of the class you used
 	 * to override the EightshiftLibs\Helpers\Component class.
 	 *
 	 * For Example: namespace\\SomeSubNamespace\\MyComponents.
 	 *
-	 * @var string
+	 * @since 1.4.0
+	 *
+	 * @var string Defaults to empty string.
 	 */
-	public $overriddenClass = '';
+	public string $overriddenClass = '';
+
+	/**
+	 * List of allowed methods that won't trigger the EscapeOutput error.
+	 *
+	 * @since 2.0.0
+	 *
+	 * @var array<string, string>
+	 */
+	public array $allowedMethods = [
+		'render' => 'render',
+		'outputCssVariables' => 'outputCssVariables',
+	];
 
 	/**
 	 * Processes this test, when one of its tokens is encountered.
+	 *
+	 * @since 1.4.0
 	 *
 	 * @param int $stackPtr The position of the current token in the stack.
 	 *
@@ -80,7 +96,7 @@ class ComponentsEscapeSniff extends EscapeOutputSniff
 
 		if ($tokens[$stackPtr]['code'] === \T_ECHO) {
 			// Check the next token after echo.
-			$elementPtr = $phpcsFile->findNext(\T_WHITESPACE, ($stackPtr + 1), null, true, null, false);
+			$elementPtr = $phpcsFile->findNext(Tokens::$emptyTokens, ($stackPtr + 1), null, true);
 
 			// If it's not the string token, move on, we're only interested in Components string.
 			if ($tokens[$elementPtr]['code'] !== \T_STRING) {
@@ -96,7 +112,7 @@ class ComponentsEscapeSniff extends EscapeOutputSniff
 			}
 
 			// Check for Components string token.
-			$componentsClassNamePtr = $phpcsFile->findNext(\T_STRING, ($stackPtr + 1), null, false, 'Components', false);
+			$componentsClassNamePtr = $phpcsFile->findNext(\T_STRING, ($stackPtr + 1), null, false, 'Components');
 
 			if (!$componentsClassNamePtr) {
 				// If there is no Components down the line, just run the regular sniff.
@@ -142,10 +158,7 @@ class ComponentsEscapeSniff extends EscapeOutputSniff
 							true
 						);
 
-						if (
-							$tokens[$methodNamePtr]['content'] === 'render' ||
-							$tokens[$methodNamePtr]['content'] === 'outputCssVariables'
-						) {
+						if (\in_array($tokens[$methodNamePtr]['content'], $this->allowedMethods, true)) {
 							return; // Skip sniffing allowed methods.
 						} else {
 							// Not allowed method, continue as usual.
@@ -178,10 +191,7 @@ class ComponentsEscapeSniff extends EscapeOutputSniff
 							true
 						);
 
-						if (
-							$tokens[$methodNamePtr]['content'] === 'render' ||
-							$tokens[$methodNamePtr]['content'] === 'outputCssVariables'
-						) {
+						if (\in_array($tokens[$methodNamePtr]['content'], $this->allowedMethods, true)) {
 							return; // Skip sniffing allowed methods.
 						} else {
 							// Not allowed method, continue as usual.
@@ -211,10 +221,7 @@ class ComponentsEscapeSniff extends EscapeOutputSniff
 						true
 					);
 
-					if (
-						$tokens[$methodNamePtr]['content'] === 'render' ||
-						$tokens[$methodNamePtr]['content'] === 'outputCssVariables'
-					) {
+					if (\in_array($tokens[$methodNamePtr]['content'], $this->allowedMethods, true)) {
 						return; // Skip sniffing allowed methods.
 					} else {
 						// Not allowed method, continue as usual.
@@ -235,7 +242,9 @@ class ComponentsEscapeSniff extends EscapeOutputSniff
 	}
 
 	/**
-	 * Checks if the import statement exists in the current file, for the given stack pointer
+	 * Checks if the import statement exists in the current file, for the given stack pointer.
+	 *
+	 * @since 1.4.0
 	 *
 	 * @param int $stackPtr The position of the current token in the stack.
 	 * @param String[] $importData Import data array.
@@ -281,11 +290,13 @@ class ComponentsEscapeSniff extends EscapeOutputSniff
 	}
 
 	/**
-	 * Return the position of the previous echo pointer
+	 * Return the position of the previous echo pointer.
+	 *
+	 * @since 1.4.0
 	 *
 	 * @param int $stackPtr The position of the current token in the stack.
 	 *
-	 * @return int Token pointer number.
+	 * @return int Echo token pointer number.
 	 */
 	private function getEchoToken(int $stackPtr): int
 	{
