@@ -5,7 +5,7 @@
  *
  * @package EightshiftCS
  *
- * @author  Eightshift <team@eightshift.com>
+ * @author  Eightshift <team.wordpress@infinum.com>
  * @license https://opensource.org/licenses/MIT MIT
  * @link    https://github.com/infinum/eightshift-coding-standards
  */
@@ -28,9 +28,9 @@ class ComponentsEscapeUnitTest extends AbstractSniffUnitTest
 	 *
 	 * @param string $testFile The name of the file being tested.
 	 *
-	 * @return array <int line number> => <int number of errors>
+	 * @return array<int, int> Key is the line number, value is the number of expected errors.
 	 */
-	public function getErrorList($testFile = ''): array
+	public function getErrorList(string $testFile = ''): array
 	{
 		switch ($testFile) {
 			case 'ComponentsEscapeUnitTest.1.inc':
@@ -59,6 +59,10 @@ class ComponentsEscapeUnitTest extends AbstractSniffUnitTest
 					25 => 1,
 					149 => 1,
 				];
+			case 'ComponentsEscapeUnitTest.6.inc':
+				return [
+					5 => 1,
+				];
 			default:
 				return [];
 		}
@@ -67,7 +71,7 @@ class ComponentsEscapeUnitTest extends AbstractSniffUnitTest
 	/**
 	 * Returns the lines where warnings should occur.
 	 *
-	 * @return array <int line number> => <int number of warnings>
+	 * @return array<int, int> Key is the line number, value is the number of expected warnings.
 	 */
 	public function getWarningList(): array
 	{
